Refactored application packaging and modules
- Shuffled packaging to use "Package by Feature" over "Package by Layer"
- Separated feature modules
- Implemented Koin dependencies
- Introduced StatusPages plugin for error handling; replaced return guards with throws
- Replaced the JWTPrincipal with a domain-friendly object containing the user
- Added a couple of useful tests
