+++ src/main/kotlin/avans/avd/auth/LoginRequest.kt
@@ -1,0 +1,9 @@
+package avans.avd.auth
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class LoginRequest (
+    val username: String,
+    val password: String
+)
