+++ src/main/kotlin/avans/avd/auth/Security.kt
@@ -1,0 +1,29 @@
+package avans.avd.auth
+
+import io.ktor.server.application.*
+import io.ktor.server.auth.*
+import io.ktor.server.auth.jwt.*
+
+fun Application.configureSecurity(
+    jwtService: JwtService
+) {
+    authentication {
+        jwt {
+            realm = jwtService.jwtRealm
+            verifier(jwtService.jwtVerifier)
+
+            validate { credential ->
+                jwtService.customValidator(credential)
+            }
+        }
+
+        jwt("access-admin-only") {
+            realm = jwtService.jwtRealm
+            verifier(jwtService.jwtVerifier)
+
+            validate { credential ->
+                jwtService.customValidator(credential)
+            }
+        }
+    }
+}
