+++ src/main/kotlin/avans/avd/auth/UserPrincipal.kt
@@ -1,0 +1,6 @@
+package avans.avd.auth
+
+import avans.avd.users.User
+import io.ktor.server.auth.*
+
+data class UserPrincipal(val user: User): Principal
