--- src/main/kotlin/avans/avd/dto/LoginRequest.kt
@@ -1,9 +1,0 @@
-package avans.avd.dto
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class LoginRequest (
-    val username: String,
-    val password: String
-)
