--- src/main/kotlin/avans/avd/dto/UserRequest.kt
@@ -1,10 +1,0 @@
-package avans.avd.dto
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class UserRequest(
-    val username: String,
-    val password: String,
-    val email: String
-)
