+++ src/main/kotlin/avans/avd/exceptions/MissingRoleException.kt
@@ -1,0 +1,5 @@
+package avans.avd.exceptions
+
+import avans.avd.users.Role
+
+class MissingRoleException(expectedRoles: List<Role> = emptyList()) : IllegalAccessException("Missing role(s): ${expectedRoles.joinToString { it.name }}")
