+++ src/main/kotlin/avans/avd/incidents/IncidentResponse.kt
@@ -1,0 +1,29 @@
+package avans.avd.incidents
+
+import avans.avd.incidents.Incident.Companion.NEW_INCIDENT_ID
+import kotlinx.datetime.Clock
+import kotlinx.datetime.LocalDateTime
+import kotlinx.datetime.TimeZone
+import kotlinx.datetime.toLocalDateTime
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class IncidentResponse (
+    val id: Long = NEW_INCIDENT_ID,
+
+    // the username of the user who issued this Incident
+    // Q. Should this be a user id?
+    val reportedBy: Long?,
+
+    val category: String,
+    val description: String,
+
+    val latitude: Double,
+    val longitude: Double,
+
+    val priority: Priority = Priority.Low,
+    val status: Status = Status.OPEN,
+    val createdAt: LocalDateTime = Clock.System.now().toLocalDateTime(TimeZone.currentSystemDefault()),
+    val updatedAt: LocalDateTime = Clock.System.now().toLocalDateTime(TimeZone.currentSystemDefault()),
+    val completedAt: LocalDateTime? = null,
+)
