--- src/main/kotlin/avans/avd/models/User.kt
@@ -1,26 +1,0 @@
-package avans.avd.models
-
-
-data class User(
-    val id: Long = NEW_USER_ID,
-    val username: String,
-    val password: String,
-    val email: String,
-    val avatar: String = DEFAULT_IMAGE,
-    val role: Role = Role.USER,
-) {
-    companion object {
-        const val NEW_USER_ID = 0L
-        // Base64 of Kotlin Kodee mascot
-        const val DEFAULT_IMAGE = "/9j/4AAQSkZJRgABAQAAAQABAAD/2wCEAAkGBwgHBgkIBwgKCgkLDRYPDQwMDRsUFRAWIB0iIiAdHx8kKDQsJCYxJx8fLT0tMTU3Ojo6Iys/RD84QzQ5OjcBCgoKDQwNGg8PGjclHyU3Nzc3Nzc3Nzc3Nzc3Nzc3Nzc3Nzc3Nzc3Nzc3Nzc3Nzc3Nzc3Nzc3Nzc3Nzc3Nzc3N//AABEIAJ8AqQMBIgACEQEDEQH/xAAcAAEBAQEBAQEBAQAAAAAAAAAABgcFBAIDAQj/xAA+EAABAwMBBAYIBAUDBQAAAAABAAIDBAURBgcSITETQVFhcZEiMkJygaHB0RQjUrEVM2LS8Agk4VNzg6LC/8QAGgEBAAIDAQAAAAAAAAAAAAAAAAECAwQFBv/EACsRAAICAgAEAwgDAAAAAAAAAAABAgMEERIhMUEFE3EVIkKBkbHh8DRRwf/aAAwDAQACEQMRAD8Aw1ERAEREAREQBERAEREAREQBERAEREAREQBERAEREAREQBERAEREAX1uu3d7dO7244Kn0LZIblVTVNYwSQU+AIzyc49vcMcu8LRwxoZ0Ya3cxjdxwx4LpYvhsr4cblpFlHZh6K91To9sgdWWeMNeOL6ZvAHvb393koNwLXFrgQQcEHqWpkY86JcMyGtH8REWAgIi61j0/XXl/wCQzcgB9Kd4w0eHae4K8ISslwxW2Dkor92z+n6HDa+XpsesYxu58P8AlRt2ttRaa11LVtAeBkObycO0LNdiXUrc1yJaaPEiItYgIiIAiIgCIiAIiIAiIgNE2bNxaap3bUY8mj7quUzs8j3NPl3653O+QH0VMvWYS1jw9DIugU3qfSsN2Dqml3Yq0DieTZfHv7/8FIizW1QtjwzXIGJVNPNSzvgqI3RysOHNcOIX8ghlqJmwwRukkecNY0ZJWtXyw0V6Y38S0slb6s0eA4Ds7wvuz2Whs8RZRxem715XcXu+P0C4vsifma37pXhJvT+iGs3ai8kOdzFO08B7x6/AKzjYyJjY42NYxow1rRgAdwX0i7FGPXRHUEWS0FD7TWsxbnY9M9IPh6KuFB7THfnW9vY158yPssHiX8aXy+4fQiURF5YxhERAEREAREQBaTpCyW59igqJ6SGeWcOc50rA7HEjAzy5LNlpWz2tbPZXUuR0lNIRj+l3EHzyuj4WoO/Ul2LR6kTqWhit18qqWnBETHAsBOcAgHHzXLV1tCs8j3MulO0ua1u5OB1Y5O+nkoVa+ZU6rpR1y7ehD6mkacqxbNDitLQ7cEjgD1neIHzwoaqvNyqp3TTVs+8Tya8tA8AOSrqpvQ7NI2fqa0+cu8vRsS0hTao1LLPcomy0FuY2R8TuIkeT6DXDrHBxI68Y5ErYzbJxjXBPlwolnLsOmteXumFTaae6SU5HoyunMTXD+kucAfgvi+R630w5ovLa+la44bJJh7HHsD+LSeHLOVo2v9s9XbL1PatMU1N0VG8xS1E7C7ee04IYAQAAQRnjnq7+ts52iw7QPxOm9UW+l6eWElu4D0dQ0cxunJDhz4HtPDHHSV1i6Sf1K7MRbq6+tx/vifGJh+i9LNb3lvrOgf70Q+i/LXunjpbVdfaWuc6GJ4dC93N0bhvN8Tg4PeCtE2QbP7VUWd+rdVNjfQs3nQQzfy91nrSP7RkEAHhwOc8FkWVevjf1J2yGZr26Ajego3D3HD/6XoZtAqR/Mt8LvdeR91qcm27SdNMaOltFY+hYdwPjija0t7QwkcPHHgudtI0PYdQ6VdrDRjYWFkZnlZA3dZNGPWO77L24JPLkcjKus/JXxjbJzTuqqa8ymnfEaepxlrC7eDx3Ht7lN7SZM3Wlj/TT583H7LkaTBOo6DdJB6Xq7MFUWs7Hc7peWy0VKZI2wtZvb7WjOSes963ndbk4b2tvfYnqiGRdSr07d6NhfPQSho5lmHgeO7lctcicJQepLRUIiKoCIiAIiIAujYbtNZrgypi9JvqyM/W3sXORWhNwkpR6oG0W6vpLpSiekkbJG7g4dbT2EdS5Fdoy0Vby9kclM48T0DgB5EED4LNqKuqqCYTUc74ZO1p5+I61R02vLnGAJ4aabA9YtLSfI4+S7UfEKLo6vjzL7T6nc1lAy36RjpIi4sY+ONpdzOMn6Lu/6bbnT093vFsleGzVkUckQPtdGXZA78Pz8Cp/X7ny6co5HNwXTMc7HIEsd91B2+tqrbWw1tBO+CpgeHxyMOC0hafievP0umkRLqUm0rStx0zqatFVBJ+EnnfJTVOPQka4kgZ/UOsd3Zgqt2CaTuNRqaHUU0EkNBRsf0cj24Ez3NLMN7QA4kkdgXWsO3sCkbDqOzmWUDBmpHAB/ix3Lz+AX4XvbJWajqILLYmMskNZKyCS41Em8+JriASMcGYzzyfEc1zypH7Z7rDdtoVxfTOa+KmDKcPHtFo9LydvD4LT7rHJc/8ATxALQC/coYTI2PmRG8dL5FrifAqPrthGoY5z/D7jbammPFsr3ujJHeMH5Erq2C/Vmx25R6e1DPDcbZVxioDqNxc6lcSQcA4y0kcviOOQQMTX+h9icctv2X3SrugMdC+SeaPpORiEYDneGWu8iv5IdjFbObvI63B5O+WZlYCf+zwB8N3ipPajtUprzbTp7S0TobZgNln3ej6Ro5MY32W8B3nlgDmBA6Kbvamoh2F5/wDRy0etvVsoZDHVVsMcjebM5cPEBQGgGb2omH9ETz8sfVeTWDt/Utcex4Hk0BdbGyJY2JxxW9y/wsnpGnUNzobhn8FVRTEcSGu4j4c1O6z03DU0stwoowypiBdI1o4SN6zjt/dQFHUy0dVFU07yyWNwc0hbWMPaN4cHDiCt2i6OfVKE1zRK5mHIv0qWCOoljacta8gHuBX5rzrWigREUAIiIAiIgCIiA0+y3C36israKqcwy9GGSwk4dke0P37ly5tnrS8mC4uDOpr4ckfEH6KEX6x1E0X8qaRnuuIXRebXZFK6vbXfei2/7K5+z6qHqV8J95hH3XndoK6j1Z6Nw9939q4DbrcWepcKtvhM4fVaXo+5G5WSJ8ry+eImOQuOSSORPwx81mxq8TJnwKLT9QtMmY9ParpKfoKWtkZCDkRQ1ha3yyAuXLpO/F7nPo3PcTku6VhJPbzV3Vw3c6hp/wAPVOZbXM35W7rTgt9nJHXkfPsXRudbFbqGarn9SJucfqPUPiVsezcd8Te1r0J4UZZJpu9R+tbpz7rd79l5nWe5tOHW6sH/AIHfZd1uvLsOcNIfFjv7l6GbQKsfzKGA+65w+65/l4T6Ta+RXke/QlgqqGWSvro3ROezcjidzwSCSR1chw8VG36YVF7rpWkOa6d+6R1jPBdq6a3rqyndBTQspWvGHPa4udjuPUpZRlW0+XGqnml3DaKPRthkudcypmYRRwOy4keu4cmj6rQr1cGWy2VFY88WN9AdrjwA81OaT1Ja6eyRU1VMKeWAEODmn0uJORjnzU3qvUDr1UhkIcyjiP5bTzcf1Fb1d1OLi7g9yf3/AAW2kjhE5OSv4iLhlAiIgCIiAIiIAiIgCIiAKg0Xdxa7qGSuxT1OGSf0n2Xf52lT6qNB2j8dcfxkzc09KQRn2n9Q+HPyWziKbvjwdSV1NKWfbRLo+WsjtrA5sUID35GN9xHD4AfuVeyzRROjbLI1hkduMDjjeOM4HkubqGxU17pt2T8udg/LmA4t7j2hejzK520uFb5/vIuzI0XtulrrLVUdDWxFh9lw4teO0FeJeUlFxepLmYwiIoAREQBERAEREAREQBERAEREAX0GPLC8NcWDgXY4BfKqNPasfQQsoa+Fk1Djd9FgBaD3cnDx496y0xhOWpy0ETcEMlRNHDCwvkkcGtaOslbDZbdHarbDRx4JYMvcPaceZXKsunbbBcv4tQSCSCRmYGcwwnmQfl3cV9a0u/8AC7UY4nYqanLGY5tHtH/O1dvEx1iVyus/V+S6WiO1neTcrsWQPP4emO5GQebut3n+wXZ01rMbrKW8uwRwZU45+99/PtUKi5Uc22NrtT6ldm1zwUtwpdyZkdRBIMjOHA94P1Unc9BQSEvttSYT/wBOUbzfPmPmpK03y4Wl3+znIjzkxP4sPw+yutM6qN5qfwktJ0cwYXl7HZbgdx4jn3rqQycXM1GyPvfvcttMi7tpq52qEzVETXQg4Mkbt4Dx6wuOtb1bUNp9O1rnDO8zowO9xx9crJFzs/Hhj2KMH2KtaCIi0SAiIgCIiAIiIAiIgCIiAIiICr0RfBbp30lZKG0cgLmuceEbh9/3wuPqO6uu91lqeIiHoRNPUwcvPn8VzEWeWTOVSqfRE7CIiwEBU+grhR0FynFZI2Iyx7rJHnDRxyQT1Z+imEWWm11WKa7BFxtAvNLUUsNDSVDJndJ0khjdkDAIAyPH5KHRFbIvlfY5yJb2ERFgICIiAIiIAiIgCIiAIiIAiIgCIiAIiIAiIgCIiAIiIAiIgCIiA//Z"
-    }
-}
-
-enum class Role {
-    USER,
-    OFFICIAL,
-    ADMIN,
-    ANONYMOUS
-}
-
-fun Role.isQualifiedOfficial() = this == Role.OFFICIAL || this == Role.ADMIN
