--- src/main/kotlin/avans/avd/plugins/Serialization.kt
@@ -1,13 +1,0 @@
-package avans.avd.plugins
-
-import io.ktor.serialization.kotlinx.json.*
-import io.ktor.server.application.*
-import io.ktor.server.plugins.contentnegotiation.*
-import io.ktor.server.response.*
-import io.ktor.server.routing.*
-
-fun Application.configureSerialization() {
-    install(ContentNegotiation) {
-        json()
-    }
-}
