--- src/main/kotlin/avans/avd/repositories/CrudRepository.kt
@@ -1,10 +1,0 @@
-package avans.avd.repositories
-
-interface CrudRepository<T, ID> {
-    suspend fun findAll(): List<T>
-    suspend fun findById(id: ID): T?
-    suspend fun save(entity: T): T
-    suspend fun saveAll(entities: Iterable<T>): List<T>
-    suspend fun delete(entity: T): Boolean
-//    suspend fun deleteAll()
-}
