--- src/main/kotlin/avans/avd/repositories/UserRepository.kt
@@ -1,7 +1,0 @@
-package avans.avd.repositories
-
-import avans.avd.models.User
-
-interface UserRepository<ID>: CrudRepository<User, ID> {
-    suspend fun findByUsername(username: String): User?
-}
