+++ src/main/kotlin/avans/avd/users/UserRepository.kt
@@ -1,0 +1,7 @@
+package avans.avd.users
+
+import avans.avd.CrudRepository
+
+interface UserRepository<ID>: CrudRepository<User, ID> {
+    suspend fun findByUsername(username: String): User?
+}
