+++ src/main/kotlin/avans/avd/users/UserRequest.kt
@@ -1,0 +1,10 @@
+package avans.avd.users
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class UserRequest(
+    val username: String,
+    val password: String,
+    val email: String
+)
