+++ src/main/kotlin/avans/avd/users/UserResponse.kt
@@ -1,0 +1,11 @@
+package avans.avd.users
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class UserResponse(
+    val id: String,
+    val username: String,
+    val email: String,
+    val role: Role
+)
