+++ src/main/resources/application.conf
@@ -4,7 +4,12 @@ ktor {
         port = ${?PORT}
     }
     application {
-        modules = [ avans.avd.ApplicationKt.module ]
+        modules = [
+            avans.avd.ApplicationKt.module,
+            avans.avd.auth.AuthModuleKt.authModule,
+            avans.avd.incidents.IncidentsModuleKt.incidentsModule,
+            avans.avd.users.UsersModuleKt.usersModule,
+        ]
     }
 }
 jwt {
